backend-python/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
click==8.3.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
fastapi==0.115.0
fonttools==4.62.0
fpdf2==2.8.1
gotrue==2.12.4
groq==1.1.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
hyperframe==6.1.0
idna==3.11
jiter==0.13.0
packaging==26.0
pillow==12.1.1
postgrest==0.16.11
pydantic>=2.11,<3
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.3
realtime==2.5.3
six==1.17.0
sniffio==1.3.1
starlette==0.38.6
storage3==0.7.7
StrEnum==0.4.15
supabase==2.7.4
supafunc==0.5.1
tqdm==4.67.3
typing_extensions==4.15.0
uvicorn==0.30.6
uvloop==0.22.1; platform_system != "Windows"
watchfiles==1.1.1
websockets==15.0.1
pdfplumber==0.11.4
numpy==1.26.4
